==> notebooks/02_transform_raw_data_into_ts_data.ipynb <==
import pandas as pd

rides = pd.read_parquet('../data/transformed/validated_rides_2022-01.parquet')

rides.head(20)

rides['pickup_hour'] = rides['pickup_datetime'].dt.floor('H')
rides

agg_rides = rides.groupby(['pickup_hour','pickup_location_id']).size().reset_index()
agg_rides.rename(columns={0:'rides'}, inplace=True)
agg_rides

from tqdm import tqdm

def add_missing_slots(agg_rides: pd.DataFrame) -> pd.DataFrame:

    location_ids = agg_rides['pickup_location_id'].unique()

    full_range = pd.date_range(
        agg_rides['pickup_hour'].min(), agg_rides['pickup_hour'].max(), freq='H')
    
    output = pd.DataFrame()

    # loop through all the unique IDs in the `pickup_location_id` columns
    # this loop adds each unique locaiton_id's pickup_hour and rides to the output dataframe
    for location_id in tqdm(location_ids):

        # keep only pickup_hour and rides for this location_id
        agg_rides_i = agg_rides.loc[agg_rides.pickup_location_id == location_id, ['pickup_hour', 'rides']]

        # quick way to add missing dates with 0 in a pandas series
        # set pickup_hour as the index
        agg_rides_i.set_index('pickup_hour', inplace=True)
        # change the index to DateTime dtype
        agg_rides_i.index = pd.DatetimeIndex(agg_rides_i.index)
        # reindex agg_rides_i to align with the list of hours in full_range; empty hours will be assigned a value of 0
        agg_rides_i = agg_rides_i.reindex(full_range, fill_value=0)

        # add back location_id columns
        agg_rides_i['pickup_location_id'] = location_id

        # concats location_id to the output dataframe defined above
        output = pd.concat([output, agg_rides_i])

    # move the purchase_day from the index to a dataframe column
    output = output.reset_index().rename(columns={'index':'pickup_hour'})

    return output

agg_rides_all_slots = add_missing_slots(agg_rides)
agg_rides_all_slots

import plotly.express as px

def plot_rides(
        rides: pd.DataFrame,
        locations = None
):
    rides_to_plot = rides[rides.pickup_location_id.isin(locations)] if locations else rides

    print(rides_to_plot)

    fig = px.line(
        rides_to_plot, 
        x='pickup_hour',
        y='rides',
        color='pickup_location_id',
        template='none',
    )

    fig.show()

plot_rides(agg_rides_all_slots, locations=[43])

agg_rides_all_slots.to_parquet('../data/transformed/ts_data_2022_01.parquet')
