==> notebooks/04_transform_raw_data_into_features_and_targets.ipynb <==
%reload_ext autoreload
%autoreload 2

Added `%reload_ext autoreload` and `%autoreload 2` above to help automatically reload modules - ensuring that changes are picked up without requiring a kernel restart.

import sys
import os

# Get the absolute path to the project root
project_root = os.path.abspath(os.path.join(os.getcwd(), '..'))
src_path = os.path.join(project_root, 'src')
sys.path.append(project_root)  # Add the project root to the Python path

# Now you can import your custom modules
from src.data import load_raw_data

# Use the load_raw_data function
rides = load_raw_data(year=2022)
print(rides)


from src.data import transform_raw_data_into_ts_data

ts_data = transform_raw_data_into_ts_data(rides)
ts_data

from src.data import transform_ts_data_into_features_and_target

features, targets = transform_ts_data_into_features_and_target(
    ts_data, 
    input_seq_len = 24*28*1,
    step_size=24
)

print(f'{features.shape}')
print(f'{targets.shape}')

tabular_data = features
tabular_data['target_rides_next_hour'] = targets

from src.paths import TRANSFORMED_DATA_DIR
tabular_data.to_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
